SystemGui/requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
click==6.7
selenium==3.9.0
joblib
python-docx